+++ CHANGELOG.md
@@ -29,12 +29,11 @@
 ## [0.1.3] - 2020-11-08
 > Update focused on quicklook and resolving flutter compile issues.
 
-- 2 major error causing building-logic is fixed.
+- 3 major error causing building-logic is fixed.
 - now there are know known compile issues.
-- the remaining issue is related to console, a flutter web runtime error, which cannot be resolved this time.
+- stack runtime error resolved.
 - quicklook has major performance boost, up to 5 times faster loading time.
 - dart character escape is implemented
-
 
 
 ## [0.1.2] - 2020-10-30